Initial commit of the project
This commit includes simple PC application and Arduino (Leonardo) code
for operating school bells. The code so far is capable of tracking time
schedule and activate/deactivate a school bell.  Operating the bell is
actually activating Arduino's GPIO with should be connected to the relay
of the school bell.

The logics of the app is working properly but the UI needs some work.
Also some bug fixing should be performed. This task will be performed in
next commits.